+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+* Added placeholders and 'not' check.
+* Added multi story support
+
 ## 0.1.1
 * Bot now reacts under the emojis, wrong reaction protection